requirements.txt
argcomplete
graphics.py
z3-solver
